fix: offset byte issue with chunked uploads (#127)
When working with the `koci` library, we found that resumable uploads
were causing a:

```
invalid: 416 Requested Range Not Satisfiable. Text: "{"errors":\[{"code":"RANGE_INVALID","message":"invalid content range"}]}
```

The `blob` we were working with was over the `chunk` limit, so when a
resumable upload was triggered, we were hitting this. We found that
setting `offset` was not being properly set and missing a `+ 1` to it's
value. But it was happening in the `send()`.

So once a second `PATCH` started, the `offset` was off by `1`, and the
error occurred.

The test we updated did not catch this since `10mb` did not cause a
second `PATCH` to fire off, thus this path was not getting triggered. We
updated it to `11mb` which exasperated the issue.

Thanks to @Racer159 and @sgettys for finding the byte issues!